rename and simplify funcs
