+++ src/state/oauth-auxiliary-resource/reducer.js
@@ -1,5 +1,5 @@
 import { combineReducers } from 'redux';
-import { restartReducer } from '../restart-utils/reducer';
+import { restartReducer } from '../restart-base/reducer';
 
 export default combineReducers({
   restartRequest: restartReducer('OAUTH_AUXILIARY_RESOURCE'),
