+++ src/state/oauth-auxiliary-resource/sagas.js
@@ -2,7 +2,7 @@ import { all } from 'redux-saga/effects';
 
 import actionCreators from './action-creators';
 import { restartOAuthAuxiliaryResource } from '../../api/oauth-auxiliary-resource';
-import { restartSagaFactory } from '../restart-utils/sagas';
+import { restartSagaFactory } from '../restart-base/sagas';
 
 const { restartWatch } = restartSagaFactory(
   'OAUTH_AUXILIARY_RESOURCE',
