+++ src/state/restart-base/reducer.js
@@ -1,0 +1,4 @@
+import { makeRequestReducerWithSubscriptionRefresh } from '../state-utils/request';
+
+export const restartReducer = (actionPrefix) =>
+  makeRequestReducerWithSubscriptionRefresh(`${actionPrefix}_RESTART`);
