--- src/state/restart-utils/reducer.js
@@ -1,5 +1,0 @@
-import { makeRequestReducerWithSubscriptionRefresh } from '../state-utils/request';
-
-export const restartReducer = (actionPrefix) => {
-  return makeRequestReducerWithSubscriptionRefresh(`${actionPrefix}_RESTART`);
-};
