+++ src/state/start-base/reducer.js
@@ -1,0 +1,4 @@
+import { makeRequestReducerWithSubscriptionRefresh } from '../state-utils/request';
+
+export const startReducer = (actionPrefix) =>
+  makeRequestReducerWithSubscriptionRefresh(`${actionPrefix}_START`);
