--- src/state/start-utils/action-creators.js
@@ -1,28 +1,0 @@
-import { makeActionCreator } from '../state-utils/action-creators';
-
-export const startActions = (actionPrefix, ...argNames) => {
-  return {
-    /**
-     * Action creator for requesting a start
-     * @param {Array} argNames Array of arguments for start request
-     */
-    startRequest: makeActionCreator(
-      `${actionPrefix}_START_REQUEST`,
-      ...argNames
-    ),
-
-    /**
-     * Action creator for marking a start as complete
-     */
-    startConfirm: makeActionCreator(
-      `${actionPrefix}_START_COMPLETE`,
-      'payload'
-    ),
-
-    /**
-     * Action creator for marking a start as failed
-     * @param {string} error The error message
-     */
-    startFail: makeActionCreator(`${actionPrefix}_START_FAIL`, 'error'),
-  };
-};
