--- src/state/start-utils/reducer.js
@@ -1,5 +1,0 @@
-import { makeRequestReducerWithSubscriptionRefresh } from '../state-utils/request';
-
-export const startReducer = (actionPrefix) => {
-  return makeRequestReducerWithSubscriptionRefresh(`${actionPrefix}_START`);
-};
