+++ src/state/stop-base/reducer.js
@@ -1,0 +1,4 @@
+import { makeRequestReducerWithSubscriptionRefresh } from '../state-utils/request';
+
+export const stopReducer = (actionPrefix) =>
+  makeRequestReducerWithSubscriptionRefresh(`${actionPrefix}_STOP`);
