--- src/state/stop-utils/reducer.js
@@ -1,5 +1,0 @@
-import { makeRequestReducerWithSubscriptionRefresh } from '../state-utils/request';
-
-export const stopReducer = (actionPrefix) => {
-  return makeRequestReducerWithSubscriptionRefresh(`${actionPrefix}_STOP`);
-};
